+++ src/Ortographic.cpp
@@ -10,7 +10,7 @@ void Ortographic::renderScene(World &world, RGBColor *frame, int startX, int sta
 	Ray ray;
 	ray.d = Vector3(0.0, 0.0, -1.0);
 	const ViewPlane &vp = world.viewPlane();
-	const double zw = 100.0; // hard-coded
+	const float zw = 100.0; // hard-coded
 
 	const int numSamples = vp.samplerState().numSamples();
 	for (int r = startY; r < startY + tileHeight; r++)
@@ -21,8 +21,8 @@ void Ortographic::renderScene(World &world, RGBColor *frame, int startX, int sta
 			for (int j = 0; j < numSamples; j++)
 			{
 				const Vector2 sp = vp.samplerState().sampleUnitSquare();
-				const double x = vp.pixelSize() * (c - 0.5 * vp.width() + sp.x);
-				const double y = vp.pixelSize() * (r - 0.5 * vp.height() + sp.y);
+				const float x = vp.pixelSize() * (c - 0.5 * vp.width() + sp.x);
+				const float y = vp.pixelSize() * (r - 0.5 * vp.height() + sp.y);
 				ray.o = Vector3(x, y, zw);
 				pixel += world.tracer().traceRay(ray, 0);
 			}
